+++ AutoTravel/Travel/WorldHelpers/Regions.cs
@@ -4,14 +4,14 @@ namespace AutoTravel.Travel.WorldHelpers
 {
     internal class Regions
     {
-        private readonly Dictionary<byte, Region> regions = [];
+        private readonly Dictionary<uint, Region> regions = [];
 
-        internal Region? Find(byte id)
+        internal Region? Find(uint id)
         {
             return this.regions.GetValueOrDefault(id);
         }
 
-        internal void Register(byte regionId, List<DataCenter> dataCenters)
+        internal void Register(uint regionId, List<DataCenter> dataCenters)
         {
             this.regions[regionId] = new Region(regionId, dataCenters);
         }
